+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.5.1] - 2025-12-04
+
+### Added
+- Add `Importmap` methods to set local imports using `package.json`
+- Add static `Importmap.importFromFile()` to import from a local JSON file
+
+### Changed
+- Update package & module `exports`
+
+## [v1.5.0] - 2025-12-03
+
+### Added
+- Add `Importmap` class with `Map`-like methods and SRI/`<script>` generating methods
+
+### Changed
+- Update `@aegisjsproject/idb` and `yaml`
 
 ## [v1.4.88] - 2025-11-19